Comments added, vars renamed
