+++ packages/lib/src/components/internal/SecuredFields/SFP/defaultProps.ts
@@ -1,46 +1,3 @@
-import { Language } from '../../../../language/Language';
-import { CardBrandsConfiguration } from '../../../Card/types';
-import { StylesObject } from '../lib/types';
-
-/**
- * Should be the only props that can be sent to SFP (from CardInput, SecuredFieldsInput, AchInput, GiftcardComponent)
- */
-export interface SFPProps {
-    allowedDOMAccess?: boolean;
-    autoFocus?: boolean;
-    brands?: string[];
-    brandsConfiguration?: CardBrandsConfiguration;
-    clientKey: string;
-    countryCode?: string;
-    hasKoreanFields?: boolean;
-    i18n: Language;
-    implementationType?: string;
-    isCollatingErrors?: boolean;
-    keypadFix?: boolean;
-    koreanAuthenticationRequired?: boolean;
-    legacyInputMode?: boolean;
-    loadingContext: string;
-    minimumExpiryDate?: string;
-    onAdditionalSFConfig?: () => {};
-    onAdditionalSFRemoved?: () => {};
-    onAllValid?: () => {};
-    onAutoComplete?: () => {};
-    onBinValue?: () => {};
-    onBrand?: () => {};
-    onChange: () => {};
-    onConfigSuccess?: () => {};
-    onError?: () => {};
-    onFieldValid?: () => {};
-    onFocus?: () => {};
-    onLoad?: () => {};
-    rootNode: HTMLElement; // Specific to SecuredFieldsInput
-    showWarnings?: boolean;
-    styles?: StylesObject;
-    trimTrailingSeparator?: boolean;
-    type: string;
-    render: () => {};
-}
-
 export default {
     type: 'card',
